lambda/requirements.txt
discord-webhook
discordwebhook
pillow
boto3 
requests
aws-xray-sdk